requirements.txt
django
psycopg2-binary
python-dotenv
gunicorn
flake8
djangorestframework

